backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
python-decouple==3.8
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
cryptography==41.0.7
requests==2.31.0
python-dateutil==2.8.2
